+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.2.2
 - fixed duplicated audio
 - removed warnings from messages
+- risk it all show correct probability
 
 # 1.2.1
 - update for appv2