+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Removed warnings from messages
 - Risk It All shows correct probability
 - Style improvements
+- Counter will show to everyone
+- support for "Phoenix Feather"
+- delay for result of avoid death increased
 
 # 1.2.1
 - update for appv2