+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 1.2.3
+- Death Move Automation Included: Add Scar and Update HP/Stress
+
 # 1.2.2
 - Fixed duplicated audio
 - Removed warnings from messages